Make delete have popup check
Delete now has popup

Fixed a few errors

+++ src/components/CourseDisplay.vue
@@ -25,7 +25,13 @@ export default {
   },
   methods: {
     deleteCourse(){
-      CourseServices.deleteCourse(this.course.courseNo);
+      // from https://www.vuemastery.com/courses/touring-vue-router/in-component-route-guards
+      const answer = window.confirm(
+        'Sure you wanna delete ' + this.course.name + '? It\'ll be gone forever!'
+      )
+      if (answer) {
+        CourseServices.deleteCourse(this.course.courseNo); // <-- Confirm delete
+      } 
     }
   }
 }
